perf: `unordered_map` の異種キーlookupを導入し `string_view` 検索時のヒープ確保を除去
`TransparentStringHash/Equal` を共通ヘッダに抽出し、`IPCHandler` / `AsyncQueryExecutor` / `AsyncConnectionExecutor` / `ConnectionRegistry` の4つの `unordered_map` に適用。
`find/contains` 計15箇所で `std::string` 一時生成を削減。
