+++ backend/database/result_cache.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "../utils/transparent_hash.h"
 #include "driver_interface.h"
 
 #include <functional>
@@ -11,16 +12,6 @@
 #include <unordered_map>
 
 namespace velocitydb {
-
-struct TransparentStringHash {
-    using is_transparent = void;
-    size_t operator()(std::string_view sv) const noexcept { return std::hash<std::string_view>{}(sv); }
-};
-
-struct TransparentStringEqual {
-    using is_transparent = void;
-    bool operator()(std::string_view a, std::string_view b) const noexcept { return a == b; }
-};
 
 struct CachedResult {
     ResultSet data;
